+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - Add `online_cpus` field to `CpuStats`
 - Fix `Image::history` response deserialization
 - Fix `Container::logs` and `Service::logs` endpoints
+<<<<<<< Updated upstream
 - Fix `Stats` field name from `network_stats` -> `networks`
+- Add missing clone implementations to some image api data types
 
 # 0.6.0
 - `name` field of `ContainerCreateOpts` is now private. Use the `ContainerCreateOpts::builder` function that takes in a `name` parameter.